% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAEM.R
\name{GAEM}
\alias{GAEM}
\title{GAEM: Genetic Algorithm Based Expectation Maximization For Inferring Gene Regulatory Networks From Incomplete Data}
\usage{
GAEM(incom, pSize, lamda, mode, sco, pcross, pmut, maxiter = 50, error = 1e-04)
}
\arguments{
\item{incom}{Incomelte data set}

\item{pSize}{Population size}

\item{mode}{Algorithm for directing the edges c("pc", "hc"),}

\item{sco}{Score measure for structure building c("loglik-g","aic-g","bic-g","pred-loglik-g","bge"),}

\item{pcross}{Probability of crossover}

\item{pmut}{Probability of mutation}

\item{maxiter}{Maximum iteration for convergence}

\item{error}{Error tollerance}

\item{lambda}{????}
}
\value{
A list containing: An imputed data set (Best), a learnt GRN (BestDag),Maximum fitness value (MaxFit) and Vector of fitness values (FitVec).
}
\description{
GAEM: Genetic Algorithm Based Expectation Maximization For Inferring Gene Regulatory Networks From Incomplete Data
}
